Add files via upload
This lab is based on Udacity's self driving car simulator, which is a nice testbed for training autonomous car using Convolutional Neural Networks.

Simulated Self Driving Car Project Demo

The simulator, written in Unity, allows you to drive a car around a track and record a video of the front view of the ride as well as the input commands.

You will upload this training data (the images + the input commands) to the GCP clusters and train a CNN to predict the steering angle command from the front view image alone. Thus, the input of your CNN is an image and the output is the steering angle.

You will then download back the trained model onto your lab machine so that you can run the car simulation in autonomous mode.